Fix: ConverterNotFoundException
tuple로 받은 뒤 dto로 변환

+++ relayRun/src/main/java/com/example/relayRun/record/dto/GetClubCalender.java
@@ -9,10 +9,10 @@ public class GetClubCalender {
 //    @JsonFormat(pattern="yyyy-MM-dd")
 //    LocalDate date;
 
-    float totalTime;
+    Double totalTime;
 
-    float totalDist;
+    Double totalDist;
 
-    float avgPace;
+    Double avgPace;
 
 }
